Updated XML declaration to match EAD3 schema
The first step is to update the XML declaration in `<ead>` to
match the EAD3 undeprecated schema.  Changing the `xmlns` attribute
to `http://ead3.archivists.org/schema` sets the document's namespace
(target schema) to EAD3; the `/undeprecated/` extension sets the schema
as undeprecated EAD3.

Since we're migrating to undeprecated EAD3 at first, we'll use
`http://ead3.archivists.org/schema/undeprecated/` as our `xmlns` value.
"Full" EAD3 is also valid under the undeprecated schema, so it isn't
strictly necessary to change this when implementing the complete standard.