Migrated to <control> and <recordid>
EAD3 replaced <eadheader> with <control>, which serves
a similar purpose*; the mapping from one to another is direct.
For purposes of generating valid EAD3, the most notable difference
is that <control> checks its attribute values against a controlled
list (whereas <eadheader> used freetext values).  Hence, we need
to change the value of scriptencoding to iso15924.

<eadid> is replaced by <recordid> in EAD3 as a required <control>
child element.  This element no longer contains information on
the country or maintaining agency (this is broken out into the new
<maintenanceagency> element), so those attributes are removed.
url becomes instanceurl, better reflecting the attribute's nature.
The value remains the same - a unique identifier for the EAD record
itself.

* More information on the change can be found in the Preface to
the EAD3 Tag Library.