Added <representation> to <control>
EAD3 introduces the <representation> element, which encodes
the location and type of other (non-EAD) forms of the finding aid.
NYPL's Archives Portal generates HTML and PDF versions of finding
aids; these URLs are provided in <representation> href attributes,
with corresponding MIME types in the linkrole attribute.  The element
value is left blank; however, it can be used for freetext descriptions
of the representation (e.g. "HTML version of finding aid").