Added three new <control> elements
EAD3 introduces three new elements to the <control> header -
<maintenancestatus>, <publicationstatus>, and <maintenanceagency>.

<maintenancestatus> "records the current version status of the EAD
instance" via one of a number of controlled values.  Since NYPL's
EAD records are generated on-the-fly by our Archives Portal, we use
'derived' as our <maintenancestatus> value.

<publicationstatus> allows for the encoding of the current publication
status of an EAD record.  Since this finding aid has been completed and
published, the value is set to 'published'.

<maintenanceagency> records identifying information about the organization
that created and/or maintains the EAD record.  <agencycode> encodes the
agency's ISIL code; <agencyname> encodes the freetext name of the agency.

In addition to these changes, EAD3 no longer supports <date> as a child
element of <titleproper>.  Encoded information on the dates of the
collection are encoded elsewhere in the record.