<languagedeclaration>, <conventiondeclaration>
Two new elements for encoding information about the finding aid -
<languagedeclaration> and <conventiondeclaration> - were added to EAD3.

<languagedeclaration> encodes the language and script of the EAD3 finding
aid in its required <language> and <script> children.  <language> encodes
language using ISO639-2b codes in the langcode attribute and optionally
freetext in the value; <script> encodes the character script using ISO15924
codes in the scriptcode attribute and optionally freetext in the value.  Note
that this is not the language/script of materials described in the finding aid,
but the language of the finding aid itself.

<conventiondeclaration> encodes information about standards used throughout
the finding aid.  In our sample document, we use the <citation> child element
to encode citations for DACS (descriptive standard used in the creation of
the finding aid) and ISO8601 (encoding standard for date metadata).