Replaced <profiledesc> with <maintenancehistory>
EAD3 removed <profiledesc> (information about the finding aid's creation
and revisions) from the standard, replacing it with the now-mandatory
<maintenancehistory>.  This element provides more structured information
about the agents, date, and events occuring to a EAD instance.

For our sample record, we encoded the creation, revision, and machine
generation of the EAD record in <maintenanceevent> elements.  The
<eventtype> element encodes the nature of the event (creation, revision,
generation); the <eventdatetime>, the  freetext and machine-readable date;
<agenttype>, whether the agent was a human or machine; and <agent>, the
name of whomever was involved.