Implemented <unitdatestructured> in <archdesc>
One of EAD3's most radical changes is the encoding of date metadata.  <unitdate>,
which contained an unstructured string as its value, was replaced by the complex
<unitdatestructured> element.  <unitdatestructured> is designed as a more
machine-readable method of encoding date metadata, and is thus more specific as
to what it contains.

Here, we replace our collection-level <archdesc> <unitdate> elements with two
<unitdatestructured> elements.  The 'type' attribute has been replaced with
'unitdatetype', with a controlled value of 'inclusive' and 'bulk'; since we have
both types of dates, we use both respectively.

Each date element represents a span of time, so we use <daterange> to represent
this.  <daterange> has two children - <fromdate>, the span's beginning, and
<todate>, its end.  Each element has a single date as its value; in addition,
the 'normal' attribute (machine-readable date form) has been replaced by
'standarddate'.