Added <physdescset> for collection-level paper material
EAD3's most extreme departure from EAD2002 is the replacement of <physdesc>
with <physdescset>.  Historically, <physdesc> was a combination of freetext
description with mixed-content elements like <extent>, resulting in data
that was difficult to transform and act upon.  EAD3 replaces this with
more powerful machine-actionable data in the form of <physdescset> and
<physdescstructured>.

<physdescset> represents a collection of data about a contiguous set of
materials.  For example, all of our physical metadata about the collection's
paper material is held in one <physdescset>.  The 'parallel' attribute
states whether each of the element's child <physdescstructured> elements
represent the same or different material.  The 'coverage' attribute states
whether the element describes the collection's material in full or in part.
Since our updated element describes the paper-based material in the
collection, 'parallel' is "true" and 'coverage' is "part".

Under <physdescset> are any number of <physdescstructured> elements.
Each one describes the material via required <quantity> and
<unittype> elements.  <physdescstructured> has two required attributes
- 'coverage', which states whether the element describes the full or
partial material *encompassed within the parent <physdescset>*, and
'physdescstructuredtype', which states the type of metadata being
provided (e.g. a count of carriers, a description of space, or
a type of material).  For both <physdescstructured> elements in our
sample, we use "whole" for 'coverage' (as it describes the full extent
of the paper material in the collection).  One element describes
the space occupied by the material whereas the other describes the
number of carriers; as such, their 'physdescstructuredtype' values
reflect this.

Note that the interplay between coverage attributes can become
complex, and that the EAD3 tag library discusses this more in depth.

<physdescstructured> has two required children - <quantity> and
<unittype>.  <quantity> represents the "how much", whereas
<unittype> represents the "what".  In our sample, we have one
<physdescstructured> for linear feet, and one for number of
boxes - their child elements are populated accordingly.