Added digital and audio physical description
EAD3's <physdescset> and <physdescstructured> paradigms also allow for
more powerful description of non-paper materials.  In our sample finding
aid, we replace EAD2002's <physdesc> with structured data using
<physdescset> and <physdescstructured>.

Similar to how we described paper materials, this collection's
electronic records are structured using a <physdescset> element.  This
joins two <physdescstructured> elements - one describing the space occupied
(file sizes) and one describing the carriers (files count).

This collection also includes audio material, which are described in EAD2002
as "19 audio files".  Since this is a simple description (and not compound
like the paper and digital physical descriptions), we do not need to use
<physdescset> to enclose parallel <physdescstructured> elements.  As such,
we will add this metadata to a <physdescstructured> element directly within
the collection-level <did>.  The 'coverage' attribute will be "part" - this
represents part of the collection - and 'physdescstructuredtype' will be
"carrier" - this represents carriers containing archival content.

Our <abstract> element will not be changed from EAD2002 - the element
is identical in EAD3 - so we have completed transforming the collection-
level <did>.