Converted Series 1-3 to EAD3
Having completed the conversion of collection-level metadata, we move on to
converting components to EAD3.  As Series 1-3 of this collection are entirely
paper-based and lightly described, the conversion is straightforward - the most
notable differences from EAD2002 are <unitdatestructured>,
<physdescstructured>, and the replacement of the 'type' attribute with
'localtype'.

<unitdatestructured> replaces each instance of <unitdate>; depending on the
type of date, it contains a <daterange>, <datesingle>, or <dateset>.  Note
that the certainty attribute is applied at the <unitdatestructured> level,
and not within its children.

Given that these series' extents are only defined at the series level,
their <physdescstructured> elements are as simple as EAD3 permits.  Each
consists of a single <unittype> and <quantity> pair, representing the series'
carriers.